refinements
- moving reusable assertions to common
- adding use memo to computed values.
