+++ libs/common/src/assertions/assertNumericString/index.ts
@@ -1,0 +1,1 @@
+export { default } from './assertNumericString';
