+++ libs/common/src/assertions/index.ts
@@ -17,3 +17,6 @@ export { default as isSessionIdLike } from './isSessionIdLike';
 export { default as assertSessionId } from './assertSessionId';
 export { default as isValidSessionId } from './isValidSessionId';
 export { default as isValidSessionKey } from './isValidSessionKey';
+export { default as isNumber } from './isNumber';
+export { default as isNumericString } from './isNumericString';
+export { default as assertNumericString } from './assertNumericString';
