+++ libs/common/src/assertions/isNumber/index.ts
@@ -1,0 +1,1 @@
+export { default } from './isNumber';
