+++ libs/common/src/assertions/isNumber/isNumber.ts
@@ -1,0 +1,5 @@
+function isNumber(value: unknown): value is number {
+  return typeof value === 'number' && Number.isFinite(value);
+}
+
+export default isNumber;
