+++ libs/common/src/assertions/isNumericString/index.ts
@@ -1,0 +1,1 @@
+export { default } from './isNumericString';
